requirements.txt
aiofiles>=24.1.0
aiohttp>=3.11.18
asyncpg>=0.30.0
azwrap>=0.1.4
click>=8.2.0
httpx>=0.28.1
pydantic>=2.11.4
pydapper>=0.12.0
python-dotenv>=1.1.0
pyyaml>=6.0.2
rich>=14.0.0
Office365-REST-Python-Client>=2.5.7
